requirements.txt
streamlit
gradio
openai
chromadb
pandas
python-dotenv
langchain
requests
